remove unnecessary headers, fix includes
It's not good practice to add separate paths under include/ to the
compiler invocation, since this can easily lead to difficult to
debug name conflicts. For example, bpf_helpers.h is present twice
in the include directory. Which one of the files gets picked up first
depends on the order of -I flags to the compiler. It's also not necessary
to include the current directory, since the compiler does this implicitly.