bpf: specialise for IP version
There is an important optimization to be had: since we know the IP
version of a socket we can attach a socket filter that is specialised
to IPv4 or IPv6. This is a big win for IPv4, since we end up hashing
a lot less data.

We also generate less code, so we are closer to not needing a higher
optmem limit for IPv4.

Fixes #2
Updates #13