bpf: use two distinct hash functions
Varying the seed to fasthash doesn't buy us additional collision resistance.
Consider two distinct buffers with the same hash:

    fasthash(A, 0) == fasthash(B, 0)

Changing the seed doesn't resolve the collision:

    fasthash(A, 42) == fasthash(B, 42)

On the other hand, using a distinct hash function it's likely that

    hashlittle(A, 0) != hashlittle(B, 0)

Due to the properties of the count-min sketch this leads to a better
estimate.