reduce IPv6 filter size below BPF_MAXINSNS
To be able to run without any privileges we need to fit the filters into
BPF_MAXINSNS aka 4096 instructions. Before this change we're at

    $ llvm-objdump --section=socket/ipv6 -S rake_bpfel.o | tail -n 2
    ;   return process_packet(skb, ETH_P_IPV6, bpf_ktime_get_ns(), bpf_get_prandom_u32(), NULL);
        4792:   95 00 00 00 00 00 00 00   exit
    $ llvm-objdump --section=socket/ipv4 -S rake_bpfel.o | tail -n 2
        2519:   bf 90 00 00 00 00 00 00   r0 = r9
        2520:   95 00 00 00 00 00 00 00   exit

Now the programs are significantly smaller:

    $ llvm-objdump --section=socket/ipv6 -S rake_bpfel.o | tail -n 2
        2294:  bf 90 00 00 00 00 00 00  r0 = r9
        2295:  95 00 00 00 00 00 00 00  exit
    $ llvm-objdump --section=socket/ipv4 -S rake_bpfel.o | tail -n 2
        2012:  bf 90 00 00 00 00 00 00  r0 = r9
        2013:  95 00 00 00 00 00 00 00  exit

The main insight enabling this is that we waste a lot of time /
instructions hashing the same bytes over and over. Instead we can
has the individual parts, and then combine the hashes as required.

Another nice side effect is that the filter is quite a bit faster:

    name              old time/op  new time/op  delta
    Rakelimit/IPv4-8   533ns ± 1%   385ns ± 3%  -27.75%  (p=0.008 n=5+5)
    Rakelimit/IPv6-8   883ns ± 3%   417ns ± 3%  -52.82%  (p=0.008 n=5+5)